requirements.txt
asgiref==3.2.7
autopep8==1.5.1
Django==3.0.5
djangorestframework==3.11.0
pycodestyle==2.5.0
pytz==2019.3
RXPY==0.1.0
sqlparse==0.3.1
